ITB-197: added health actuator, configured cors and openapi listener
