+++ ecs.yaml
@@ -118,7 +118,7 @@ Resources:
     Type: AWS::ElasticLoadBalancingV2::TargetGroup
     Properties:
       HealthCheckIntervalSeconds: 180
-      HealthCheckPath: '/user?page-size=1&page=1'
+      HealthCheckPath: '/actuator/health'
       HealthCheckTimeoutSeconds: 30
       HealthCheckPort: '8080'
       UnhealthyThresholdCount: 2
@@ -142,6 +142,7 @@ Resources:
             - /user/*
             - /login
             - /login/*
+            - /v3/api-docs/accountservice
       Actions:
         - Type: forward
           TargetGroupArn: !Ref TargetGroup
