+++ src/main/java/com/inthebytes/accountservice/AccountserviceApplication.java
@@ -1,15 +1,36 @@
 package com.inthebytes.accountservice;
 
+import com.inthebytes.accountservice.login.JwtProperties;
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 @SpringBootApplication
-@ComponentScan(basePackages={"com.inthebytes.accountservice"})
+@OpenAPIDefinition(info =
+	@Info(title = "Account Service API", version = "1.0", description = "Documentation Account Service API v1.0")
+)
 public class AccountserviceApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(AccountserviceApplication.class, args);
+	}
+
+	@Bean
+	public WebMvcConfigurer corsConfigurer() {
+		return new WebMvcConfigurer() {
+			@Override
+			public void addCorsMappings(CorsRegistry registry) {
+				registry.addMapping("/**")
+					.allowedOriginPatterns("*")
+					.exposedHeaders(JwtProperties.HEADER_STRING)
+					.allowCredentials(true)
+					.allowedMethods("HEAD", "GET", "PUT", "POST", "DELETE", "PATCH");
+			}
+		};
 	}
 
 }
