+++ src/main/resources/application.properties
@@ -7,3 +7,6 @@ spring.datasource.name = stacklunch
 
 ## Hibernate Properties
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
+
+# Open API URL
+springdoc.api-docs.path = /v3/api-docs/accountservice
